Moved awsRum into App instead of index
